Character w/ light set up.
It works, a little bit janky but it works. sometimes the light will tick down even when the player isnt holding the light up. and i cant seem to find a way to set the light to fade whilst the item isnt held up. but for now this will do.